Bug #1441864 bsp: camera: Pclint scan
[root cause  ] value type inconsistent
[changes     ] bsp:camera
[side effects] NA
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]:s9863a1h10
[test case]:camera function
[reviewers   ] yangyang.liu
[change_type ] feature_bugfix
[tag_product ] common

Fix Pclint scan wran.

+++ core/isp2.6/src/cam_trusty.c
@@ -588,7 +588,7 @@ bool cam_trusty_isp_pitch_set(uint32_t y_pitch, uint32_t u_pitch,
 	return 1;
 }
 
-bool cam_trusty_isp_3dnr_fetch_set(unsigned long chroma, uint32_t luma,
+bool cam_trusty_isp_3dnr_fetch_set(unsigned long chroma, unsigned long luma,
 	uint32_t pitch)
 {
 	pr_info("no trusty version\n");
